+++ src/Serilog.Sinks.Kafka.TestApp/Program.cs
@@ -36,7 +36,7 @@ namespace Serilog.Sinks.Kafka.TestApp
 
             while (true)
             {
-                Thread.Sleep(10);
+                Thread.Sleep(1000);
                 Log.Information("Console Application Test!");
             }
