+++ src/Serilog.Sinks.Kafka.TestApp/appsettings.json
@@ -14,8 +14,8 @@
       {
         "Name": "Kafka",
         "Args": {
-          "batchSizeLimit": "50",
-          "period": "5",
+          "batchSizeLimit": "2",
+          "period": "2",
           "bootstrapServers": "47.96.43.6:9092",
           "topic": "logs"
         }
